divided schema into many files
