+++ pkg/openrocket/schema_common.go
@@ -1,0 +1,32 @@
+package openrocket
+
+import (
+	"encoding/xml"
+	"fmt"
+)
+
+// Contains the 'utility' or repeated schema
+
+// AxialOffset represents the axial offset element of the XML document
+type AxialOffset struct {
+	XMLName xml.Name `xml:"axialoffset"`
+	Method  string   `xml:"method,attr"`
+	Value   float64  `xml:",chardata"`
+}
+
+// String returns full string representation of the AxialOffset
+func (a *AxialOffset) String() string {
+	return fmt.Sprintf("AxialOffset{Method=%s, Value=%.2f}", a.Method, a.Value)
+}
+
+// Position represents the position element of the XML document
+type Position struct {
+	XMLName xml.Name `xml:"position"`
+	Value   float64  `xml:",chardata"`
+	Type    string   `xml:"type,attr"`
+}
+
+// String returns full string representation of the Position
+func (p *Position) String() string {
+	return fmt.Sprintf("Position{Value=%.2f, Type=%s}", p.Value, p.Type)
+}
